requirements.txt
gensim==3.8.1
inspyred==1.0.1
joblib==0.14.1
matplotlib==3.2.1
nltk==3.4.5
numpy==1.22.3
pandas==1.0.3
pyclust==0.2.0
pykalman==0.9.5
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
statsmodels==0.11.1
treelib==1.6.1
Unidecode==1.1.1
Cython
seaborn
keras-tcn
